Crate will_fuzzy

Will’s Fuzzy Algorithm Library

for linux (?) primarily. not tested on other devices lol

will_fuzzy is a quite fast, low runtime allocation high efficienty fuzzy matching engine and psudo-TUI framework.

This library provides the core search algorithms and session management required to build custom fuzzy finders. additionally, by following the trait patterns defined, custom algorithms could be inserted as long as they match the required functions! it additionally defines structures to allow for any struct to implement a fuzzy search across it, allowing weighting of different fields as well!

Workflow

1. Define an entity list to be searched against

Either use GenericStringStruct + CandidateGenerator::from_lines(Vec<String>)

Or implement the FuzzyCandidate trait for a custom struct:

fn search_targets(&self) -> Vec<ScoreTarget>; (use ScoreTarget::new())

fn display_text(&self) -> &str;

2. Pick an algorithm

Currently, only AlgoWillGreedyVer2 is supported.

let matcher = FuzzyMatcher::with_algo(AlgoWillGreedyVer2::new());

3. Generate a session

let session = SearchSession::create(&entities, matcher, String::new());

4. Interface with the session

You can interact with the session directly via:

.top_results(amt, offset)

.set_query(str)

.type_char(c) — Primarily for use in interactive forms!

5. Use the TUI Interface

Alternatively, use the built-in TUI framework:

let mut fuzzy_app = FuzzyApp::new(session);
fuzzy_app.init()?;

Re-exports

pub use fuzzy::canidate::FuzzyCandidate;

pub use fuzzy::canidate::ScoreTarget;

pub use fuzzy::canidate::ScoredResult;

pub use fuzzy::session::SearchSession;

pub use fuzzy::matcher::FuzzyMatcher;

pub use fuzzy::algorithms::algo_greedy_v2::AlgoWillGreedyVer2;

pub use entities::DesktopEntity;

pub use entities::GenericStringStruct;

pub use entities::AnimalEnt;

Modules

app

mostly self contained. just create via new, and init it with a session!

entities

Entities Module

fuzzy

Fuzzy Algorithms, Session Engine

Enums

SearchMode

technically usable on non-ascii, i haven’t tested it extensively. so. just incase