will_fuzzy

Module entities

Entities Module

the entities module does some pre-defined implementations of structs with the FuzzyCanidate traits. custom ones can be done via the trait pre-defined:

DesktopEntity: parses/handles Linux .desktop files for application launching (theoredically).

parse_desktop_inis()

AnimalEnt: a simple entity model used for demonstration and scale testing, but left in for fun (yay!)

GenericStringStruct: A versatile standard struct for basic line strings, with a helper to pass it any file and reading by lines

CandidateGenerator: nice way to convert a Vec into a basic GenericStringStruct type.

animal_demo: for animals

new_generic_from_file: lets you pass a file. see the main demo

parse_desktop_inis: reads ini files. it kinda works.

Re-exports

pub use animal::AnimalEnt;

pub use animal::animal_demo;

pub use desktop::DesktopEntity;

pub use desktop::parse_desktop_inis;

pub use generic::GenericStringStruct;

pub use generic::new_generic_from_file;

pub use generic::CandidateGenerator;

Modules

animal

desktop

generic

prelude