will_fuzzy::entities

Module prelude

Re-exports

pub use crate::fuzzy::canidate::FuzzyCandidate;

pub use crate::fuzzy::canidate::ScoreTarget;

pub use crate::SearchSession;

pub use crate::fuzzy::algorithms::*;

Modules

fmt

Utilities for formatting and printing Strings.

Structs

BufReader

The BufReader<R> struct adds buffering to any reader.

File

An object providing access to an open file on the filesystem.

Path

A slice of a path (akin to str).

Enums

ErrorKind

A list specifying general categories of I/O error.

Traits

BufRead

A BufRead is a type of Reader which has an internal buffer, allowing it to perform extra ways of reading.