will_fuzzy

Module fuzzy

Fuzzy Algorithms, Session Engine

The fuzzy module is the heart & soul of the library’s search capabilities,

algoritms!

matching logic!

session state stuff!

FuzzyCandidate: if implemented, allows for passing to a session!

SearchSession: state controller(?) manager(?) the thing handling the algorithm.

history and current query management

letter by letter parsing

flat string querying

results polling

interfaces for the tui or other handlers if designed!

threshold for multi character type culling for optimization

Re-exports

pub use self::algorithms::SimilarityAlgorithm;

pub use self::canidate::ScoreTarget;

pub use self::canidate::*;

pub use self::session::*;

pub use self::matcher::*;

pub use self::algorithms::*;

Modules

algorithm

algorithms

canidate

matcher

session

Functions

scale_weight

the way weights are handled, to optimize calcs, is that its ints scaled as a 1024 and bit shifted. so be sure to use handlers!