will_fuzzy::fuzzy::algorithms::algo_greedy_v2

Struct AlgoWillGreedyVer2

pub struct AlgoWillGreedyVer2 {
    pub bonus_bound: i64,
    pub bonus_consec: i64,
    pub bonus_start: i64,
}

Remade Algorithm, might actually do what it says we traverse down the query, and see if the letter maps to somewhere within the target. if so, check for 1 of three conditonals it could be we define 3 core weights mappins to three conditions:

boundary: on match, if the previous is a deliminator -, _, or ' '

consecutive: if the previous matched index also was found, so multiple in a row

start: if the first letter of query matches the first of target

Fields
bonus_bound: i64bonus_consec: i64bonus_start: i64
Implementations

impl AlgoWillGreedyVer2

pub fn new() -> Self

pub fn with_weights( custom_bonus_bound: i64, custom_bonus_consecutive: i64, custom_bonus_start: i64, ) -> Self

pub fn modify_weights( &mut self, custom_bonus_bound: i64, custom_bonus_consecutive: i64, custom_bonus_start: i64, )

pub fn one_step_calc(&self, target: &[u8], query: &[u8]) -> i64

core philosopy is fast fail, fast iter, minimal allocations