will_fuzzy::fuzzy::canidate

Trait FuzzyCandidate

pub trait FuzzyCandidate {
    // Required methods
    fn search_targets(&self) -> Vec<ScoreTarget<'_>>;
    fn display_text(&self) -> &str;

    // Provided methods
    fn exec(&self) -> String { ... }
    fn usage_bonus(&self) -> i64 { ... }
    fn display_candidate(&self) -> String { ... }
}

using a trait to define parts of a struct to score + weights hardcode weight 1?

Required Methods

fn search_targets(&self) -> Vec<ScoreTarget<'_>>

for structs, define which strings are included in scoring?

fn display_text(&self) -> &str

Provided Methods

fn exec(&self) -> String

fn usage_bonus(&self) -> i64

from use statistics, later include ig

fn display_candidate(&self) -> String